refactor(infer): delete receiver.rs — the CST path is the only resolution path
Move-don't-rewrite relocation of the three survivors to their sole callers:
`uppercase_dotted_type_prefix` → chain.rs; `fun_trailing_lambda_this_type` and
`resolve_call_params` → cst_lambda.rs. receiver.rs and its `mod` declaration are
gone; every deleted function verified at zero references.

Also folds in the PR #202 review finding: `cursor_node_at` looked the cursor
line up via `str::lines()`, which yields no final empty line for `\n`-terminated
content — a cursor on the trailing empty line (end-of-file, e.g. right after
typing `items.forEach {` + Enter) returned `None`, failing `it` resolution and
the brace-repair gate before they could run, and emptying the completion
lambda-scope stack. The trailing-gap cursor now resolves on the last character
of the last content line (node ends are exclusive), whose node carries the
enclosing context. Pinned by `it_resolves_on_trailing_empty_line_at_eof`
(RED without the fix, GREEN with it).
